+++ src/components/dropdown-list/scroll-mask.component.ts
@@ -1,5 +1,8 @@
 import {Component, HostListener, EventEmitter} from '@angular/core';
 
+/**
+ * The scroll mask is a transparent div covering the entire viewport which is intended to prevent scrolling.
+ */
 @Component({
     selector: 'gtx-scroll-mask',
     template: ``